+++ wolox-books/src/app/app-routing.module.ts
@@ -1,13 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-const routes: Routes = [
-    {
-        path: '',
-        pathMatch: 'prefix',
-        redirectTo: '/login',
-    }
-]
+const routes: Routes = []
 
 @NgModule({
     imports: [RouterModule.forRoot(routes)],
