--- wolox-books/src/app/screens/Sign-up/sign-up.component.scss
@@ -1,48 +1,0 @@
-@import '../../../scss/colors';
-@import '../../../scss/buttons';
-
-.container {
-  background-color: lighten($grey, 20%);
-  border-top: 5px solid $blue;
-  display: grid;
-  padding: 30px 20px;
-  width: 330px;
-}
-
-.logo {
-  justify-self: center;
-  width: 270px;
-}
-
-.title {
-  font-weight: bold;
-  justify-self: center;
-  letter-spacing: 5px;
-  margin-top: 10px;
-}
-
-.form {
-  border-bottom: 2px solid lighten($grey, 10%);
-  padding: 30px 0 20px;
-}
-
-.control {
-  display: flex;
-  flex-direction: column;
-  margin: 20px 0;
-}
-
-.ctrl-label {
-  margin: 5px 0;
-  padding: 0 10px;
-}
-
-.ctrl-input {
-  border-radius: 5px;
-  height: 35px;
-}
-
-.btn-login {
-  color: $green;
-  margin-top: 20px;
-}
