+++ wolox-books/src/app/screens/User/screens/Login/login.component.html
@@ -1,0 +1,2 @@
+<p>login works!</p>
+<button class="btn btn-default" [routerLink]="['/sign-up']">Sign up</button>
