+++ wolox-books/src/app/screens/User/screens/Login/login.module.ts
@@ -1,14 +1,14 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-
-import { LoginRoutingModule } from './login-routing.module';
 import { LoginComponent } from './login.component';
+import { RouterModule } from '@angular/router';
 
 @NgModule({
   declarations: [LoginComponent],
   imports: [
+    RouterModule,
     CommonModule,
-    LoginRoutingModule
-  ]
+  ],
+  exports: [LoginComponent]
 })
 export class LoginModule { }
