+++ wolox-books/src/app/screens/User/screens/Sign-up/sign-up.component.scss
@@ -1,0 +1,24 @@
+@import 'src/scss/colors';
+@import 'src/scss/buttons';
+
+.form {
+  border-bottom: 2px solid lighten($grey, 10%);
+  margin-bottom: 20px;
+  padding: 30px 0 20px;
+}
+
+.control {
+  display: flex;
+  flex-direction: column;
+  margin: 20px 0;
+}
+
+.ctrl-label {
+  margin: 5px 0;
+  padding: 0 10px;
+}
+
+.ctrl-input {
+  border-radius: 5px;
+  height: 35px;
+}
