+++ wolox-books/src/app/screens/User/screens/Sign-up/sign-up.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { SignUpComponent } from './sign-up.component';
 import { HttpClientModule } from '@angular/common/http';
-import { SignUpRoutingModule } from './sign-up-routing.module';
+import { RouterModule } from '@angular/router';
 
 @NgModule({
   declarations: [
@@ -12,9 +12,9 @@ import { SignUpRoutingModule } from './sign-up-routing.module';
   imports: [
     CommonModule,
     FormsModule,
+    RouterModule,
     ReactiveFormsModule,
     HttpClientModule,
-    SignUpRoutingModule
   ],
   exports: [SignUpComponent]
 })
