+++ wolox-books/src/app/screens/User/user.component.html
@@ -1,0 +1,5 @@
+<div class="container">
+  <img class="logo" src="../assets/wolox_logo.svg" alt="Wolox">
+  <h4 class="title">{{title}}</h4>
+  <router-outlet></router-outlet>
+</div>
