+++ wolox-books/src/app/screens/User/user.component.scss
@@ -1,0 +1,22 @@
+@import 'src/scss/colors';
+@import 'src/scss/buttons';
+
+.container {
+  background-color: lighten($grey, 20%);
+  border-top: 5px solid $blue;
+  display: grid;
+  padding: 30px 20px;
+  width: 330px;
+}
+
+.logo {
+  justify-self: center;
+  width: 270px;
+}
+
+.title {
+  font-weight: bold;
+  justify-self: center;
+  letter-spacing: 5px;
+  margin-top: 10px;
+}
