+++ wolox-books/src/app/screens/User/user.module.ts
@@ -1,0 +1,19 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { UserRoutingModule } from './user-routing.module';
+import { UserComponent } from './user.component';
+import { LoginModule } from './screens/Login/login.module';
+import { RegisterModule } from './screens/Sign-up/sign-up.module';
+
+@NgModule({
+  declarations: [UserComponent],
+  imports: [
+    CommonModule,
+    UserRoutingModule,
+    RegisterModule,
+    LoginModule,
+  ],
+  exports: [UserComponent]
+})
+export class UserModule { }
