--- wolox-books/src/app/screens/login/login.component.html
@@ -1,1 +1,0 @@
-<p>login works!</p>
