+++ wolox-books/src/scss/buttons.scss
@@ -12,3 +12,7 @@
   background-color: $green;
   color: $white;
 }
+
+.btn-default {
+  color: $green;
+}
